Completely remove SubRepoPermsStore from non-enterprise
